chore: move `StatusIcon` to `@/components/common`
It makes more sense there than in `@/components/layout`
