+++ services/orchest-webserver/client/src/components/common/StatusIcon.tsx
